Contact Backend

+++ Service.Catalog/Dtos/Company/ContactListDto.cs
@@ -4,7 +4,7 @@
     {
         public int IdContacto { get; set; }
         public string Nombre { get; set; }
-        public int Telefono { get; set; }
+        public long Telefono { get; set; }
         public string Correo { get; set; }
         public bool Activo { get; set; }
     }
